feat(routes/home): create home route with main poke-fetch
