+++ src/routes/Layout.js
@@ -1,0 +1,10 @@
+import React from "react";
+
+function Layout() {
+    return(
+        <div data-testid='app'>
+        <Navigation />
+        <Outlet />
+</div>
+    );
+}
